[check-credentials] add check-credentials command (#12)
Hey, I don't really know Go or how to test this locally, but I think
this should do what i'm looking for. Just trying to add a command that I
can use to understand if the credentials cache is valid or not. Maybe
there's a better way to do this already?

---------

Co-authored-by: spalger <spalger@users.noreply.github.com>
Co-authored-by: Harry Bowron <hbowron@gmail.com>